pageindex-rag/requirements.txt
# PageIndex RAG Requirements
# Base dependencies
python-dotenv
pandas
numpy
# Note: PageIndex specific dependencies should be added after reviewing its docs
